+++ src/commands/settings/prefix.ts
@@ -36,16 +36,16 @@ export default class PrefixCommand extends Command {
     // Reset the default prefix
     if (!args.prefix) {
       await provider.update(id, { settings: { prefix: DEFAULTS.PREFIX } });
-      return this.success(message, this.t('COMMAND_PREFIX_RESPONSE_RESET', message, DEFAULTS.PREFIX));
+      return this.success(message, message.t('COMMAND_PREFIX_RESPONSE_RESET', DEFAULTS.PREFIX));
     }
 
     // Check if the prefix actually changes
     const settings = provider.fetch(id)?.settings;
-    if (settings?.prefix === args.prefix) return this.warning(message, this.t('COMMAND_PREFIX_RESPONSE_IDENTICAL', message));
+    if (settings?.prefix === args.prefix) return this.warning(message, message.t('COMMAND_PREFIX_RESPONSE_IDENTICAL'));
 
     // Save the new prefix
     await provider.update(id, { settings: { prefix: args.prefix } });
-    return this.success(message, this.t('COMMAND_PREFIX_RESPONSE_MODIFIED', message, args.prefix));
+    return this.success(message, message.t('COMMAND_PREFIX_RESPONSE_MODIFIED', args.prefix));
   }
 
 }