+++ tests/suites/commands/job.ts
@@ -36,7 +36,7 @@ export const job = (client: Client) => describe('Job', () => {
     spies.embed = jest.spyOn(command, 'embed');
     spies.warning = jest.spyOn(command, 'warning');
     spies.error = jest.spyOn(command, 'error');
-    spies.t = jest.spyOn(command, 't');
+    spies.t = jest.spyOn(message, 't');
 
     if (command && rest.length) args = await command.parse(message, rest);
 
@@ -60,13 +60,13 @@ export const job = (client: Client) => describe('Job', () => {
   it('should warn if a level is provided without a job', async () => {
     await setup('!job 150');
     expect(spies.warning).toBeCalledTimes(1);
-    expect(spies.t).toHaveBeenCalledWith('COMMAND_JOBS_ARGUMENTS_LEVEL_IGNORED', message);
+    expect(spies.t).toHaveBeenCalledWith('COMMAND_JOBS_ARGUMENTS_LEVEL_IGNORED');
   });
 
   it('should warn if an unparsable argument is provided', async () => {
     await setup('!job tailor invalid');
     expect(spies.warning).toBeCalledTimes(1);
-    expect(spies.t).toHaveBeenCalledWith('COMMAND_JOBS_ARGUMENTS_UNKNOWN', message, ['invalid']);
+    expect(spies.t).toHaveBeenCalledWith('COMMAND_JOBS_ARGUMENTS_UNKNOWN', ['invalid']);
   });
 
   it('should correct the level if below 1', async () => {
@@ -87,35 +87,35 @@ export const job = (client: Client) => describe('Job', () => {
     await setup(`!job tailor 100 <@${message.member!.id}>`);
     await command.exec(message, args);
     expect(spies.error).toBeCalledTimes(1);
-    expect(spies.t).toHaveBeenCalledWith('COMMAND_JOBS_RESPONSE_INVALID_COMBINATION', message, args);
+    expect(spies.t).toHaveBeenCalledWith('COMMAND_JOBS_RESPONSE_INVALID_COMBINATION', args);
   });
 
   it('should warn if no member have the job provided', async () => {
     await setup('!job lumberjack');
     await command.exec(message, args);
     expect(spies.warning).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOBODY', message, 'Lumberjack');
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOBODY', 'COMMAND_JOB_RESPONSE_JOB_LUMBERJACK');
   });
 
   it('should warn if no jobs are registered for the member provided', async () => {
     await setup(`!job <@${message.member!.id}>`);
     await command.exec(message, args);
     expect(spies.warning).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOJOBS', message, message.member!.displayName);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOJOBS', message.member!.displayName);
   });
 
   it('should warn if no jobs are registered for the member who sent the message', async () => {
     await setup(`!job`);
     await command.exec(message, args);
     expect(spies.warning).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOJOBS', message, message.member!.displayName);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOJOBS', message.member!.displayName);
   });
 
   it('should warn if the provided job is not referenced for the provided member', async () => {
     await setup(`!job tailor <@${message.member!.id}>`);
     await command.exec(message, args);
     expect(spies.warning).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOJOB', message, message.member!.displayName, 'Tailor');
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_NOJOB', message.member!.displayName, 'COMMAND_JOB_RESPONSE_JOB_TAILOR');
   });
 
   it('should save the level to the database', async () => {
@@ -125,35 +125,35 @@ export const job = (client: Client) => describe('Job', () => {
     const id = `${message.guild!.id}:${message.member!.id}`;
     expect(spies.update).toBeCalledWith(id, { id, jobs: { tailor: 100 } });
     expect(spies.embed).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_SINGLE', message);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_SINGLE');
   });
 
   it('should display the level of all members with the job provided', async () => {
     await setup('!job tailor');
     await command.exec(message, args);
     expect(spies.embed).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_JOB_TAILOR', message);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_JOB_TAILOR');
   });
 
   it('should display all jobs for the member provided', async () => {
     await setup(`!job <@${message.member!.id}>`);
     await command.exec(message, args);
     expect(spies.embed).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_ALL', message);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_ALL');
   });
 
   it('should display all jobs for the member who sent the message', async () => {
     await setup(`!job`);
     await command.exec(message, args);
     expect(spies.embed).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_ALL', message);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_ALL');
   });
 
   it('should display the provided job for the member provided', async () => {
     await setup(`!job tailor <@${message.member!.id}>`);
     await command.exec(message, args);
     expect(spies.embed).toBeCalledTimes(1);
-    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_SINGLE', message);
+    expect(spies.t).toBeCalledWith('COMMAND_JOB_RESPONSE_TITLE_SINGLE');
   });
 
 });